Header re-do. Logo links to dashboard for dashboard layouts. Moved account page to dashboard layout
